requirements.txt
fastapi==0.115.4
imageio==2.36.0
numpy==2.0.2
Pillow==11.0.0
readmrz==0.0.2
pytesseract